refactor: move cookies into zhttp.http package
