+++ example/src/main/scala/example/CookieServerSide.scala
@@ -1,7 +1,6 @@
 package example
 
-import zhttp.http._
-import zhttp.http.cookie.Cookie
+import zhttp.http.{Cookie, _}
 import zhttp.service.Server
 import zio._
