+++ zio-http/src/main/scala/zhttp/http/CookieEncoder.scala
@@ -1,6 +1,7 @@
-package zhttp.http.cookie
+package zhttp.http
+
 import io.netty.handler.codec.http.{cookie => jCookie}
-import zhttp.http.cookie.Cookie.{Request, SameSite}
+import Cookie.{Request, SameSite}
 
 sealed trait CookieEncoder[A] {
   final def apply(a: Cookie[A]): String = unsafeEncode(a, false)
