--- zio-http/src/main/scala/zhttp/http/cookie/package.scala
@@ -1,6 +1,0 @@
-package zhttp.http
-
-package object cookie {
-  type RequestCookie  = Cookie[Cookie.Request]
-  type ResponseCookie = Cookie[Cookie.Response]
-}
