+++ zio-http/src/main/scala/zhttp/http/middleware/Csrf.scala
@@ -1,7 +1,6 @@
 package zhttp.http.middleware
 
 import zhttp.http._
-import zhttp.http.cookie.Cookie
 import zio.ZIO
 
 import java.util.UUID
